feat: add search API for organizations by name
Implemented a new API endpoint to search for organizations by their name. This allows users to query organizations based on the provided name.

+++ src/main/java/com/esd_project/repo/HrRepo.java
@@ -11,11 +11,11 @@ import java.util.List;
 
 public interface HrRepo extends JpaRepository<HR, Long> {
     @Query("SELECT h FROM HR h WHERE h.organization.id = :organizationId")
-    List<HR> findByOrganizationId(@Param("organizationId") Long organizationId);
+    List<HR> findByOrganizationId(@Param("organizationId") int organizationId);
     @Modifying
     @Transactional
     @Query("DELETE FROM HR h WHERE h.organization.id = :organizationId")
-    void deleteByOrganizationId(@Param("organizationId") Long organizationId);
+    void deleteByOrganizationId(@Param("organizationId") int organizationId);
 
     boolean existsByOrganizationId(Long organizationId);
 }
